GO-2658 add SyncDbAfterVersions fallback mechanism
